requirements.txt
diff-match-patch==20121119
Django==1.10.5
django-bootstrap3==8.2.1
django-filter==1.0.2
django-import-export==0.5.1
django-jquery==3.1.0
django-leaflet==0.20.0
django-location-field==2.0.2
django-map-widgets==0.1.8
django-osm-field==0.3.1
django-queryset-csv==1.0.0
Jinja2==2.9.5
MarkupSafe==0.23
psycopg2==2.6.2
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
tablib==0.11.4
unicodecsv==0.14.1
